+++ src/components/Web3Status/index.tsx
@@ -2,7 +2,7 @@
 import { t, Trans } from '@lingui/macro'
 import { useWeb3React } from '@web3-react/core'
 import { Connector } from '@web3-react/types'
-import { getWalletForConnector } from 'connectors'
+import { getWalletForConnector } from 'connectors/utils'
 import { darken } from 'polished'
 import { useMemo } from 'react'
 import { Activity } from 'react-feather'
